create version 0.21
git-svn-id: https://svn.code.sf.net/p/odin-win/code/trunk@24 b91d169b-bf28-450c-9b59-d70c078a1df8

+++ src/ODIN/DriveList.cpp
@@ -242,8 +242,8 @@ if (!bSuccess)
 unsigned CDriveInfo::GetDiskNumber()
 {
   unsigned no;
-  int pos1 = fDeviceName.find(L"Harddisk") + 8;
-  int pos2 = fDeviceName.find(L'\\', pos1);
+  size_t pos1 = fDeviceName.find(L"Harddisk") + 8;
+  size_t pos2 = fDeviceName.find(L'\\', pos1);
   no = _wtoi(fDeviceName.substr(pos1, pos2-pos1).c_str());
   return no;
 }
